float -> int, s'il te plait

+++ src/Controller/Vault/GetVaultStatusController.php
@@ -51,7 +51,7 @@ class GetVaultStatusController
             ->getSingleScalarResult();
 
         $isOpen = $vault->isOpen();
-        $openCost = 10 + floor($itemCount / 10);
+        $openCost = 10 + (int)floor($itemCount / 10);
 
         return $responseService->success([
             'isOpen' => $isOpen,
